server/requirements.txt
alembic==1.4.2
apispec==5.1.1
flask==1.0.2
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
SQLAlchemy==1.3.8
sqlalchemy-filters==0.8.0
python-docx==0.8.11
google-cloud-storage==1.43.0
secure-smtplib==0.1.1
pydrive==1.3.1
google_auth_oauthlib==0.4.6
zipfile38==0.0.3
py3dns==3.2.1
Flask-Mail==0.9.1
python-dateutil==2.8.2
secure-smtplib==0.1.1
itsdangerous
numpy
pytz==2022.2.1
passlib==1.7.4
bcrypt==4.0.0 
flask-apispec==0.11.4
flask_cors==3.0.10
flask_jwt_extended==4.0.0
mysqlclient

